+++ lib/client/group/group.ts
@@ -2,11 +2,12 @@ import {Group} from "@/lib/types/Group";
 import {revalidateOptions} from "@/lib/client/common";
 import {ServerSideException} from "@/lib/client/exception/ServerSideException";
 import createLogger, {Logger} from "@/lib/logger";
+import {ClientSideException} from "@/lib/client/exception/ClientSideException";
 
 async function getGroups(): Promise<Group[]> {
     const url = `http://localhost:8082/api/group`
     try {
-        const res = await fetch(url, revalidateOptions)
+        const res = await fetch(url, { next: { revalidate: 60, tags: ['groups'] } })
         if (!res.ok) return []
         return res.json()
     } catch {
@@ -21,7 +22,12 @@ async function getGroups(): Promise<Group[]> {
  */
 async function getGroupByName(groupName: string): Promise<Group> {
     const logger: Logger = createLogger("getGroupByName")
-    const url = `http://localhost:8082/api/group?name=${groupName}`
+    let url: string
+    if (groupName) {
+        url = `http://localhost:8082/api/group?name=${groupName}`;
+    } else {
+            throw new ClientSideException("groupName must not be null");
+    }
     const res: Response = await fetch(url, revalidateOptions)
     if (!res.ok) {
         logger.warn("received non-200 response")