+++ CHANGELOG.md
@@ -15,5 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - add docker build script and git hooks
 
 ### Changed
+- use revalidateTag instead of revalidatePath to prevent GroupPage re-render
 - add cursor pointer to DeleteGroupButton
 - add createGroup client method and sidebar sheet UI